// api/models/faqCategory.ts
/**
 * Generated by orval v7.19.0 🍺
 * Do not edit manually.
 * MagicMessenger
 * OpenAPI spec version: v1
 */

export type FaqCategory = (typeof FaqCategory)[keyof typeof FaqCategory];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FaqCategory = {
  General: "General",
  Privacy: "Privacy",
  Pricing: "Pricing",
} as const;